feat: make binding function public
