+++ binding.go
@@ -1,16 +1,29 @@
 package ltask
 
 import (
-	"os"
-	"path"
-	"strings"
-	"time"
-	"unsafe"
+	"sync/atomic"
 
 	"go.yuchanns.xyz/lua"
 )
 
-func ltaskOpen(L *lua.State) int {
+// OpenLibs opens ltask library.
+// This is useful when you don't write Go code and just want to use ltask directly in Lua.
+func OpenLibs(L *lua.State, lib *lua.Lib) {
+	L.GetGlobal("package")
+	_ = L.GetField(-1, "preload")
+
+	L.PushLightUserData(lib)
+
+	l := []*lua.Reg{
+		{Name: "ltask.bootstrap", Func: OpenBootstrap},
+	}
+	L.SetFuncs(l, 1)
+	L.Pop(2)
+}
+
+// OpenCore represents `require("ltask")` in Lua.
+// This is useful when you want to integrate ltask to your own lua-binding projects.
+func OpenCore(L *lua.State) int {
 	l := []*lua.Reg{
 		{Name: "pack", Func: LuaSerdePack},
 		{Name: "unpack", Func: LuaSerdeUnpack},
@@ -53,275 +66,66 @@ func ltaskOpen(L *lua.State) int {
 	return 1
 }
 
-func getS(L *lua.State) *serviceUd {
-	ud := L.ToUserData(L.UpValueIndex(1))
-	if ud == nil {
-		panic("Invalid service userdata")
+var bootInit atomic.Int32
+
+// OpenBootstrap represents `require("ltask.bootstrap")` in Lua.
+// This is useful when you want to integrate ltask to your own lua-binding projects.
+func OpenBootstrap(L *lua.State) int {
+	if bootInit.Add(1) != 1 {
+		return L.Errorf("ltask.bootstrap can only require once")
 	}
-	return (*serviceUd)(ud)
-}
+	l := []*lua.Reg{
+		{Name: "searchpath", Func: ltaskSearchPath},
+		{Name: "readfile", Func: ltaskReadFile},
+		{Name: "loadfile", Func: ltaskLoadFile},
+		{Name: "dofile", Func: ltaskDoFile},
+		{Name: "deinit", Func: ltaskDeinit},
+		{Name: "run", Func: ltaskRun},
+		{Name: "wait", Func: ltaskWait},
+		{Name: "post_message", Func: lpostMessage},
+		{Name: "new_service", Func: ltaskNewService},
+		{Name: "init_timer", Func: ltaskInitTimer},
+		{Name: "init_root", Func: ltaskInitRoot},
+		{Name: "pushlog", Func: ltaskBootPushLog},
+		// We don't need `init_socket` here, as it is proceed by Go runtime automatically.
+		{Name: "pack", Func: LuaSerdePack},
+		{Name: "unpack", Func: LuaSerdeUnpack},
+		{Name: "remove", Func: LuaSerdeRemove},
+		{Name: "unpack_remove", Func: LuaSerdeUnpackRemove},
+	}
 
-func ltaskSleep(L *lua.State) int {
-	csec := L.OptInteger(1, 0)
-	time.Sleep(centisecond * time.Duration(csec))
-	return 0
-}
+	L.NewLib(l)
 
-func ltaskEventWait(L *lua.State) int {
-	event := (*sockEvent)(L.ToUserData(L.UpValueIndex(1)))
-	r := event.wait()
-	L.PushBoolean(r > 0)
+	L.PushLightUserData(L.ToUserData(L.UpValueIndex(1)))
+	l2 := []*lua.Reg{
+		{Name: "init", Func: ltaskInit},
+	}
+	L.SetFuncs(l2, 1)
 	return 1
 }
 
-func ltaskEventInit(L *lua.State) int {
-	s := getS(L)
-	index := s.task.services.getSockevent(s.id)
-	if index >= 0 {
-		return L.Errorf("Already init event")
-	}
-	index = int64(s.task.allocSockevent())
-	if index < 0 {
-		return L.Errorf("Too many sockevents")
-	}
-	event := &s.task.event[index]
-	L.PushLightUserData(event)
-	L.PushGoClousure(ltaskEventWait, 1)
-	if !event.open() {
-		return L.Errorf("Create sockevent fail")
-	}
-	s.task.services.initSockevent(s.id, index)
-	fd := event.fd()
-	L.PushLightUserData(*(*unsafe.Pointer)(unsafe.Pointer(&fd)))
-	return 2
-}
+var rootInit atomic.Int32
 
-func ltaskTimerAdd(L *lua.State) int {
-	s := getS(L)
-	t := s.task.timer
-	if t == nil {
-		return L.Errorf("Init timer before bootstrap")
+// OpenRoot represents `require("ltask.root")` in Lua.
+// This is useful when you want to integrate ltask to your own lua-binding projects.
+func OpenRoot(L *lua.State) int {
+	if rootInit.Add(1) != 1 {
+		return L.Errorf("ltask.root can only require once")
 	}
-	ev := &timerEvent{
-		session: int32(L.CheckInteger(1)),
-		id:      s.id,
+	l := []*lua.Reg{
+		{Name: "init_service", Func: ltaskInitService},
+		{Name: "close_service", Func: ltaskCloseService},
 	}
-	ti := L.CheckInteger(2)
-	if ti < 0 || ti != int64(int32(ti)) {
-		return L.Errorf("Invalid timer time: %d", ti)
-	}
-	t.Add(ev, time.Duration(ti)*centisecond)
-	return 0
-}
 
-func ltaskTimerUpdate(L *lua.State) int {
-	s := getS(L)
-	t := s.task.timer
-	if t == nil {
-		return L.Errorf("Init timer before bootstrap")
+	L.NewLibTable(l)
+
+	if L.GetField(lua.LUA_REGISTRYINDEX, "LTASK_ID") != lua.LUA_TLIGHTUSERDATA {
+		return L.Errorf("No service id, the VM is not inited by ltask")
 	}
-	if L.GetTop() > 1 {
-		L.SetTop(1)
-		L.CheckType(1, lua.LUA_TTABLE)
-	}
-	var idx int64
-	t.Update(func(event *timerEvent) {
-		v := int64(event.session)<<32 | int64(event.id)
-		L.PushInteger(v)
-		idx++
-		L.SetI(1, int64(idx))
-	})
-	n := int64(L.RawLen(1))
-	for i := int64(idx + 1); i <= n; i++ {
-		L.PushNil()
-		L.SetI(1, i)
-	}
+	ud := (*serviceUd)(L.ToUserData(-1))
+	L.Pop(1)
+
+	L.PushLightUserData(ud)
+	L.SetFuncs(l, 1)
 	return 1
-}
-
-func ltaskNow(L *lua.State) int {
-	s := getS(L)
-	t := s.task.timer
-	if t == nil {
-		return L.Errorf("Init timer before bootstrap")
-	}
-	start := t.Start()
-	now := t.Now()
-	L.PushInteger(start + now/100)
-	L.PushInteger(start*100 + now)
-	return 2
-}
-
-func lself(L *lua.State) int {
-	s := getS(L)
-	L.PushInteger(int64(s.id))
-	return 1
-}
-
-func lworkerId(L *lua.State) int {
-	s := getS(L)
-	worker := s.task.getWorkerId(s.id)
-	if worker >= 0 {
-		L.PushInteger(int64(worker))
-		return 1
-	}
-	return 0
-}
-
-func lworkerBind(L *lua.State) int {
-	s := getS(L)
-	if L.IsNoneOrNil(1) {
-		s.task.services.setBindingThread(s.id, threadNone)
-		return 0
-	}
-	worker := L.CheckInteger(1)
-	if worker < 0 || worker >= int64(len(s.task.workers)) {
-		return L.Errorf("Invalid worker id: %d", worker)
-	}
-	s.task.services.setBindingThread(s.id, int32(worker))
-	return 0
-}
-
-func ltaskLabel(L *lua.State) int {
-	s := getS(L)
-	label := s.task.services.getLabel(s.id)
-	L.PushString(label)
-	return 1
-}
-
-func ltaskPushLog(L *lua.State) int {
-	L.CheckType(1, lua.LUA_TLIGHTUSERDATA)
-	data := L.ToUserData(1)
-	sz := L.CheckInteger(2)
-	s := getS(L)
-	if !s.task.pushLog(s.id, data, sz) {
-		return L.Errorf("log error")
-	}
-
-	return 0
-}
-
-func ltaskPopLog(L *lua.State) int {
-	s := getS(L)
-	m, ok := s.task.lqueue.pop()
-	if !ok {
-		return 0
-	}
-	start := int64(0)
-	t := s.task.timer
-	if t != nil {
-		start = t.Start() * 100
-	}
-	L.PushInteger(m.timestamp + start)
-	L.PushInteger(int64(m.id))
-	L.PushLightUserData(m.msg)
-	L.PushInteger(m.sz)
-	return 4
-}
-
-func ltaskReadFile(L *lua.State) int {
-	fileName := L.CheckString(1)
-	var (
-		err   error
-		scode []byte
-	)
-	for _, embedfs := range embedfsList {
-		scode, err = embedfs.ReadFile(fileName)
-		if err == nil {
-			L.PushString(string(scode))
-			return 1
-		}
-		if !os.IsNotExist(err) {
-			break
-		}
-	}
-	L.PushNil()
-	L.PushString(err.Error())
-	return 2
-}
-
-func ltaskDoFile(L *lua.State) int {
-	fileName := L.CheckString(1)
-	var (
-		err   error
-		scode []byte
-	)
-	for _, embedfs := range embedfsList {
-		scode, err = embedfs.ReadFile(fileName)
-		if err == nil {
-			L.SetTop(0)
-			err = L.DoString(string(scode))
-			if err != nil {
-				L.PushNil()
-				L.PushString(err.Error())
-				return 2
-			}
-			return L.GetTop()
-		}
-		if !os.IsNotExist(err) {
-			break
-		}
-	}
-	L.PushNil()
-	L.PushString(err.Error())
-	return 2
-}
-
-func ltaskLoadFile(L *lua.State) int {
-	fileName := L.CheckString(1)
-	var (
-		err   error
-		scode []byte
-	)
-	for _, embedfs := range embedfsList {
-		scode, err = embedfs.ReadFile(fileName)
-		if err == nil {
-			err = L.LoadString(string(scode))
-			if err != nil {
-				L.PushNil()
-				L.PushString(err.Error())
-				return 2
-			}
-			return 1
-		}
-		if !os.IsNotExist(err) {
-			break
-		}
-	}
-	L.PushNil()
-	L.PushString(err.Error())
-	return 2
-}
-
-func ltaskSearchPath(L *lua.State) int {
-	name := L.CheckString(1)
-	pattern := L.CheckString(2)
-
-	patterns := strings.Split(pattern, ";")
-
-	modulePath := strings.ReplaceAll(name, ".", "/")
-
-	for _, pat := range patterns {
-		fullPattern := strings.ReplaceAll(pat, "?", modulePath)
-
-		dir := path.Dir(fullPattern)
-		base := path.Base(fullPattern)
-
-		for _, embedfs := range embedfsList {
-			entries, err := embedfs.ReadDir(dir)
-			if err != nil {
-				continue
-			}
-			for _, entry := range entries {
-				if entry.IsDir() {
-					continue
-				}
-				if entry.Name() == base {
-					L.PushString(path.Join(dir, entry.Name()))
-					return 1
-				}
-			}
-		}
-	}
-	return 0
 }
