+++ root.go
@@ -6,30 +6,6 @@ import (
 	"go.yuchanns.xyz/lua"
 )
 
-var rootInit atomic.Int32
-
-func ltaskRootOpen(L *lua.State) int {
-	if rootInit.Add(1) != 1 {
-		return L.Errorf("ltask.root can only require once")
-	}
-	l := []*lua.Reg{
-		{Name: "init_service", Func: ltaskInitService},
-		{Name: "close_service", Func: ltaskCloseService},
-	}
-
-	L.NewLibTable(l)
-
-	if L.GetField(lua.LUA_REGISTRYINDEX, "LTASK_ID") != lua.LUA_TLIGHTUSERDATA {
-		return L.Errorf("No service id, the VM is not inited by ltask")
-	}
-	ud := (*serviceUd)(L.ToUserData(-1))
-	L.Pop(1)
-
-	L.PushLightUserData(ud)
-	L.SetFuncs(l, 1)
-	return 1
-}
-
 func ltaskInitService(L *lua.State) int {
 	s := getS(L)
 	sid := L.CheckInteger(1)
